Add locales middleware and collocate static paths
